Fixed small bug for tabs. A lot of renaming. Fixed rgba order on sliders for edit spaces. Smoothing out edges on value parameter system. Added enum iteration. Added realtime listeners to parameter sets - and fixed a couple of local/global parameter handle inconsistencies.

+++ infrastructure/parameters/CustomFormatters.h
@@ -151,7 +151,7 @@ namespace cpl
 		void setValues(std::vector<std::string> valuesToConsume) 
 		{
 			this->values = std::move(valuesToConsume); 
-			transformer.setQuantization(values.size()); 
+			transformer.setQuantization(static_cast<int>(values.size()));
 		}
 		const std::vector<std::string> & getValues() const noexcept { return values; }
